Move user errors to presenter
