+++ user-rest-service/interfaces/presenter/user_errors.go
@@ -1,18 +1,18 @@
-package userdomain
+package presenter
 
 import (
 	"encoding/json"
 	"log"
 )
 
-type ValidationError struct {
+type UserValidationError struct {
 	UserID   string `json:"user_id,omitempty"`
 	Name     string `json:"name,omitempty"`
 	Email    string `json:"email,omitempty"`
 	Password string `json:"password,omitempty"`
 }
 
-func (e *ValidationError) Error() string {
+func (e *UserValidationError) Error() string {
 	b, err := json.Marshal(e)
 	if err != nil {
 		log.Println(err)
@@ -21,7 +21,7 @@ func (e *ValidationError) Error() string {
 	return string(b)
 }
 
-func (e *ValidationError) IsEmpty() bool {
+func (e *UserValidationError) IsEmpty() bool {
 	if e.UserID != "" ||
 		e.Name != "" ||
 		e.Email != "" ||
@@ -32,12 +32,12 @@ func (e *ValidationError) IsEmpty() bool {
 	return true
 }
 
-type ConflictError struct {
+type UserConflictError struct {
 	UserID string `json:"user_id,omitempty"`
 	Email  string `json:"email,omitempty"`
 }
 
-func (e *ConflictError) Error() string {
+func (e *UserConflictError) Error() string {
 	b, err := json.Marshal(e)
 	if err != nil {
 		log.Println(err)
