+++ user-rest-service/router/router.go
@@ -14,11 +14,11 @@ import (
 	"github.com/rs/cors"
 
 	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/config"
-	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/handler"
 	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/infrastructure/externalapi"
 	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/infrastructure/externalapi/client"
 	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/infrastructure/persistence"
 	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/infrastructure/persistence/db"
+	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/interfaces/handler"
 	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/usecase"
 )
